Deep_Reflective_Reader/requirements.txt
openai>=1.0.0
llama-index>=0.10.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
numpy>=1.24.0
dependency-injector>=4.49.0